Check configs exist before attempting to delete.

+++ src/lib/response-triggers/roleconfig.ts
@@ -80,7 +80,11 @@ export default class RoleConfig extends Trigger {
 					break;
 				}
 
-				// TODO check config exists first
+				if (!this.config.getRoles(words[2]).includes(words[3] as Role)) {
+					newMessage = `Role "${words[3]}" is not set for user "${words[2]}"`;
+					break;
+				}
+
 				this.config.deleteRole(words[2], words[3] as Role);
 				newMessage = 'Role has been deleted';
 				break;
